Code cleanup

+++ src/libc/mem.c
@@ -21,7 +21,7 @@ void memcpy(uint8_t *src, uint8_t *dst, int nbytes) {
  * @param val           Value to set.
  * @param len           Length of memory portion.
  */
-void memset(uint8_t *dst, uint8_t val, uint32_t len) {
+void memset(uint8_t *dst, uint8_t val, size_t len) {
     uint8_t *temp = (uint8_t *)dst;
     for (; len != 0; --len) *temp++ = val;
 }
@@ -33,7 +33,7 @@ uint32_t free_mem_addr = 0x10000;
  * @param align         Pages should be aligned (4K)?
  * @param physical_addr Physical base address of the allocated memory (buffer).
  */
-uint32_t kmalloc(uint32_t size, int align, uint32_t *physical_addr) {
+uint32_t kmalloc(size_t size, int align, uint32_t *physical_addr) {
     if (align == 1 && (free_mem_addr & 0xfffff000)) {
         free_mem_addr &= 0xfffff000;
         free_mem_addr += 0x1000; // 4K
