+++ src/libc/mem.h
@@ -5,7 +5,8 @@
 #ifndef MEM_H
 #define MEM_H
 
-#include "../cpu/types.h"
+#include <stddef.h>
+#include <stdint.h>
 
 /* Copy a portion of memory from a source to a destination.
  * @param src           Source address.
@@ -19,13 +20,13 @@ void memcpy(uint8_t *src, uint8_t *dst, int nbytes);
  * @param val           Value to set.
  * @param len           Length of memory portion.
  */
-void memset(uint8_t *dst, uint8_t val, uint32_t len);
+void memset(uint8_t *dst, uint8_t val, size_t len);
 
 /* Memory allocator.
  * @param size          Size of the memory chunk needed.
  * @param align         Pages should be aligned (4K)?
  * @param physical_addr Physical base address of the allocated memory (buffer).
  */
-uint32_t kmalloc(uint32_t size, int align, uint32_t *physical_addr);
+uint32_t kmalloc(size_t size, int align, uint32_t *physical_addr);
 
 #endif
